refactor!: move the main command from root to run
