+++ cmd/run_test.go
@@ -1,0 +1,112 @@
+/*
+ * Copyright 2025 - 2026 Zigflow authors <https://github.com/mrsimonemms/zigflow/graphs/contributors>
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package cmd
+
+import (
+	"errors"
+	"fmt"
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+)
+
+func TestPanicMessage(t *testing.T) {
+	tests := []struct {
+		Name     string
+		Input    any
+		Expected string
+	}{
+		{
+			Name:     "error value",
+			Input:    errors.New("something went wrong"),
+			Expected: "something went wrong",
+		},
+		{
+			Name:     "string value",
+			Input:    "a plain string",
+			Expected: "a plain string",
+		},
+		{
+			Name:     "other value",
+			Input:    42,
+			Expected: fmt.Sprintf("%+v", 42),
+		},
+	}
+
+	for _, test := range tests {
+		t.Run(test.Name, func(t *testing.T) {
+			assert.Equal(t, test.Expected, panicMessage(test.Input))
+		})
+	}
+}
+
+func TestBuildDataConverter(t *testing.T) {
+	tests := []struct {
+		Name        string
+		ConvertData bool
+		KeyPath     string
+		ExpectNil   bool
+		ExpectError bool
+	}{
+		{
+			Name:        "disabled returns nil without reading key file",
+			ConvertData: false,
+			KeyPath:     "",
+			ExpectNil:   true,
+		},
+		{
+			Name:        "enabled with missing key file returns error",
+			ConvertData: true,
+			KeyPath:     "/nonexistent/path/keys.yaml",
+			ExpectNil:   true,
+			ExpectError: true,
+		},
+	}
+
+	for _, test := range tests {
+		t.Run(test.Name, func(t *testing.T) {
+			dc, err := buildDataConverter(test.ConvertData, test.KeyPath)
+
+			if test.ExpectError {
+				assert.Error(t, err)
+			} else {
+				assert.NoError(t, err)
+			}
+
+			if test.ExpectNil {
+				assert.Nil(t, dc)
+			} else {
+				assert.NotNil(t, dc)
+			}
+		})
+	}
+}
+
+func TestNewRunCmd_Flags(t *testing.T) {
+	cmd := newRunCmd()
+
+	assert.NotNil(t, cmd.Flags().Lookup("file"))
+	assert.NotNil(t, cmd.Flags().Lookup("validate"))
+	assert.NotNil(t, cmd.Flags().Lookup("temporal-address"))
+	assert.NotNil(t, cmd.Flags().Lookup("temporal-namespace"))
+	assert.NotNil(t, cmd.Flags().Lookup("convert-data"))
+	assert.NotNil(t, cmd.Flags().Lookup("cloudevents-config"))
+	assert.NotNil(t, cmd.Flags().Lookup("env-prefix"))
+	assert.NotNil(t, cmd.Flags().Lookup("health-listen-address"))
+	assert.NotNil(t, cmd.Flags().Lookup("metrics-listen-address"))
+	assert.NotNil(t, cmd.Flags().Lookup("disable-telemetry"))
+}
