+++ docs/docs/getting-started/your-first-workflow.mdx
@@ -31,7 +31,7 @@ do:
 You can now run this workflow with Zigflow
 
 ```sh
-zigflow -f ./workflow.yaml
+zigflow run -f ./workflow.yaml
 ```
 
 ## Triggering your workflow
