+++ examples/external-calls/compose.yaml
@@ -63,6 +63,8 @@ services:
       WORKFLOW_FILE: ./examples/external-calls/workflow.yaml
     volumes:
       - ../../:/go/app
+    command:
+      - run
     links:
       - grpc
       - temporal
